Move back to the `concept_macros` file
