+++ cudax/include/cuda/experimental/__stream/get_stream.cuh
@@ -23,7 +23,7 @@
 
 #include <cuda_runtime_api.h>
 
-#include <cuda/std/__concepts/__concept_macros.h>
+#include <cuda/std/__concepts/concept_macros.h>
 #include <cuda/std/__concepts/convertible_to.h>
 #include <cuda/std/__cuda/api_wrapper.h>
 #include <cuda/std/__type_traits/is_convertible.h>
