+++ libcudacxx/include/cuda/std/__concepts/constructible.h
@@ -20,7 +20,7 @@
 #  pragma system_header
 #endif // no system header
 
-#include <cuda/std/__concepts/__concept_macros.h>
+#include <cuda/std/__concepts/concept_macros.h>
 #include <cuda/std/__concepts/convertible_to.h>
 #include <cuda/std/__concepts/destructible.h>
 #include <cuda/std/__type_traits/add_lvalue_reference.h>
