+++ libcudacxx/include/cuda/std/__concepts/regular.h
@@ -20,7 +20,7 @@
 #  pragma system_header
 #endif // no system header
 
-#include <cuda/std/__concepts/__concept_macros.h>
+#include <cuda/std/__concepts/concept_macros.h>
 #include <cuda/std/__concepts/equality_comparable.h>
 #include <cuda/std/__concepts/semiregular.h>
