+++ libcudacxx/include/cuda/std/__concepts/relation.h
@@ -20,7 +20,7 @@
 #  pragma system_header
 #endif // no system header
 
-#include <cuda/std/__concepts/__concept_macros.h>
+#include <cuda/std/__concepts/concept_macros.h>
 #include <cuda/std/__concepts/predicate.h>
 
 _LIBCUDACXX_BEGIN_NAMESPACE_STD
