+++ libcudacxx/include/cuda/std/detail/libcxx/include/optional
@@ -169,7 +169,7 @@ template<class T>
 #  pragma system_header
 #endif // no system header
 
-#include <cuda/std/__concepts/__concept_macros.h>
+#include <cuda/std/__concepts/concept_macros.h>
 #include <cuda/std/__concepts/invocable.h>
 #include <cuda/std/__exception/terminate.h>
 #include <cuda/std/__functional/hash.h>
